Rename urdf_tactile/tactile.h -> urdf_tactile/sensor.h
