+++ urdf_tactile/test/parser.cpp
@@ -1,4 +1,4 @@
-#include "urdf_tactile/tactile.h"
+#include "urdf_tactile/sensor.h"
 #include "urdf_tactile/cast.h"
 #include <urdf/sensor.h>
 #include <fstream>
